publish: add publishing tool
This tool recursively looks into the directory pointed by `PUBLISH_SRC`
for HTML or HTM files, looks up all local `href` and `src`-bound tags,
uploads their value as a blob alongside the page content to the configured PDS.

The HTML is edited, replacing `href` and `src` attributes with the AT URI
of the uploaded value.

Needs README, may be a lil sharp.